requirements.txt
fastapi
uvicorn
pandas
scikit-learn
pyodbc
python-dotenv